PS-1956: Adjust datatype for some timers used for extended slow query log
The ulong and ulint are used for such timers as innodb_io_reads_wait_timer,
innodb_lock_que_wait_timer and innodb_innodb_que_wait_timer. They are
32-bit at some platforms which may be not enough in some cases.
Adjust corresponding datatypes to 64-bit.

+++ sql/sql_class.cc
@@ -1525,7 +1525,7 @@ THD::~THD() {
 
 extern "C" void thd_report_innodb_stat(THD *thd, unsigned long long trx_id,
                                        enum mysql_trx_stat_type type,
-                                       unsigned long long value) {
+                                       uint64_t value) {
   assert(thd);
   assert(!thd_is_background_thread(thd));
   (void)thd;
